apps/backend/requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2
psycopg[binary]>=3.2
redis>=5.0
django-environ>=0.11
celery>=5.4
django-celery-beat>=2.6
django-celery-results>=2.5
djangorestframework>=3.15
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
requests
pathlib
datetime
dnspython